Sentiment Analysis Project with NLTK (Amazon Reviews)
This project is a sentiment analysis model that uses The Natural Language Toolkit (NLTK) and machine learning techniques to classify the sentiment of Amazon product reviews. The model is trained on a dataset of 10,000 Amazon product reviews, where each review is labeled as either positive, negative, or neutral.

The project involves several steps, including data preprocessing, feature extraction, and model training. The data preprocessing step involves cleaning the raw text data by removing punctuation, stop words, and other noise. The feature extraction step involves converting the preprocessed text data into numerical features that can be used as input to a machine learning model.

The code for this project is written in Python and uses several libraries, including NLTK, Scikit-learn, Pandas, and Numpy. The project also includes detailed documentation and comments to help readers understand the code and the methodology used.